rename: study 컨트롤러 패키지 경로 변경
